pyproject.toml
[project]
name = "automaton-auditor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "gitpython>=3.1.46",
    "langchain>=1.2.10",
    "langchain-openai>=1.1.10",
    "langgraph>=1.0.9",
    "openai>=2.24.0",
    "pillow>=12.1.1",
    "pydantic>=2.12.5",
    "pypdf>=6.7.2",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "rich>=14.3.3",
    "tree-sitter>=0.25.2",
    "typing-extensions>=4.15.0",
]
